Add  the RFC 8693 `scope` claim to generated JWTs
Extend the default claims struct in the new  `rfc8693Claims` struct and
use it to build claims. It includes all the RFC 8693 claims, but
currently populate only `scope` in addition to the standard RFC 7519 JWT
claims.

Mention the new `scope` claim and RFC 8693 in the README and OpenAPI
doc, and add a test to ensure it populates from the form.

Refactor the tests to thoroughly test JWT contents including RFC 8693
claims.

While at it, properly test the default content-type in `/resource`
responses.